src/inference/requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mlflow==2.7.1
python-dotenv==1.0.0
boto3==1.28.44 